feature/SA-1 Constants file renamed api -> errors
