+++ constants/errors.ts
