+++ constants/index.ts
@@ -1,2 +1,2 @@
-export * from './api';
+export * from './errors';
 export * from './routes';
